(#232) GitLab split and __init__
